+++ vibe-core/src/main/java/org/molgenis/vibe/core/database_processing/query_string_creation/QueryStringPathRange.java
@@ -1,4 +1,4 @@
-package org.molgenis.vibe.core.tdb_processing.query_string_creation;
+package org.molgenis.vibe.core.database_processing.query_string_creation;
 
 import org.apache.jena.query.Syntax;
